Use SWTBotJunit runner to have screenshots on failure
