+++ org.moreunit.swtbot.test/src/org/moreunit/create/MethodCreationTest.java
@@ -5,15 +5,17 @@ import static org.fest.assertions.Assertions.assertThat;
 import org.eclipse.jdt.core.IMethod;
 import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.swtbot.eclipse.finder.widgets.SWTBotEclipseEditor;
+import org.eclipse.swtbot.swt.finder.junit.SWTBotJunit4ClassRunner;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.moreunit.ConditionCursorLine;
 import org.moreunit.JavaProjectSWTBotTestHelper;
 import org.moreunit.test.context.Project;
 import org.moreunit.test.context.Properties;
 import org.moreunit.test.context.TestType;
 
-
+@RunWith(SWTBotJunit4ClassRunner.class)
 public class MethodCreationTest extends JavaProjectSWTBotTestHelper
 {
 	@Project(mainSrc = "MethodCreation_class_with_method.txt",
