+++ org.moreunit.swtbot.test/src/org/moreunit/jump/BasicJumpTest.java
@@ -4,9 +4,11 @@ import static org.fest.assertions.Assertions.assertThat;
 
 import org.eclipse.swtbot.eclipse.finder.widgets.SWTBotEclipseEditor;
 import org.eclipse.swtbot.eclipse.finder.widgets.SWTBotEditor;
+import org.eclipse.swtbot.swt.finder.junit.SWTBotJunit4ClassRunner;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.moreunit.ConditionCursorLine;
 import org.moreunit.JavaProjectSWTBotTestHelper;
 import org.moreunit.test.context.Context;
@@ -15,7 +17,7 @@ import org.moreunit.test.context.Properties;
 import org.moreunit.test.context.TestType;
 import org.moreunit.test.context.configs.SimpleJUnit4Project;
 
-
+@RunWith(SWTBotJunit4ClassRunner.class)
 public class BasicJumpTest extends JavaProjectSWTBotTestHelper
 {
 	@Before
@@ -31,18 +33,14 @@ public class BasicJumpTest extends JavaProjectSWTBotTestHelper
     @Context(SimpleJUnit4Project.class)
     public void should_jump_to_test_when_shortcut_is_pressed_in_cut()
     {
-    	openResource("SomeClass.java");
-    	getShortcutStrategy().pressJumpShortcut();
-    	assertThat(bot.activeEditor().getTitle()).isEqualTo("SomeClassTest.java");
+        testSimpleJump("SomeClass.java", "SomeClassTest.java");
     }
 
 	@Test
     @Context(SimpleJUnit4Project.class)
     public void should_jump_to_cut_when_shortcut_is_pressed_in_testcase()
     {
-    	openResource("SomeClassTest.java");
-    	getShortcutStrategy().pressJumpShortcut();
-    	assertThat(bot.activeEditor().getTitle()).isEqualTo("SomeClass.java");
+	    testSimpleJump("SomeClassTest.java", "SomeClass.java");
     }
     
     @Project(
