+++ org.moreunit.swtbot.test/src/org/moreunit/refactoring/RenameClassTest.java
@@ -3,10 +3,12 @@ package org.moreunit.refactoring;
 import static org.fest.assertions.Assertions.assertThat;
 
 import org.eclipse.swtbot.eclipse.finder.waits.Conditions;
+import org.eclipse.swtbot.swt.finder.junit.SWTBotJunit4ClassRunner;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotShell;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotTreeItem;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.moreunit.JavaProjectSWTBotTestHelper;
 import org.moreunit.test.context.Preferences;
 import org.moreunit.test.context.Project;
@@ -16,6 +18,7 @@ import org.moreunit.test.context.TestType;
 @Preferences(testClassNameTemplate="${srcFile}Test", 
              testSrcFolder="test", 
              testType=TestType.JUNIT4)
+@RunWith(SWTBotJunit4ClassRunner.class)
 public class RenameClassTest extends JavaProjectSWTBotTestHelper 
 {
 	@Project(
