+++ org.moreunit.swtbot.test/src/org/moreunit/refactoring/RenameMethodTest.java
@@ -9,8 +9,10 @@ import org.eclipse.jface.dialogs.IDialogConstants;
 import org.eclipse.ltk.internal.ui.refactoring.RefactoringUIMessages;
 import org.eclipse.swtbot.eclipse.finder.waits.Conditions;
 import org.eclipse.swtbot.eclipse.finder.widgets.SWTBotEclipseEditor;
+import org.eclipse.swtbot.swt.finder.junit.SWTBotJunit4ClassRunner;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotShell;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.moreunit.ConditionCursorLine;
 import org.moreunit.JavaProjectSWTBotTestHelper;
 import org.moreunit.SWTBotHelper;
@@ -22,6 +24,7 @@ import org.moreunit.test.context.TestType;
              testSrcFolder = "test", 
              testType = TestType.JUNIT4,
              testMethodPrefix = true)
+@RunWith(SWTBotJunit4ClassRunner.class)
 public class RenameMethodTest extends JavaProjectSWTBotTestHelper
 {
 	@Project(
