+++ org.moreunit.swtbot.test/src/org/moreunit/refactoring/RenamePackageTest.java
@@ -4,10 +4,12 @@ import static org.fest.assertions.Assertions.assertThat;
 import static org.junit.Assert.fail;
 
 import org.eclipse.jdt.core.ICompilationUnit;
+import org.eclipse.swtbot.swt.finder.junit.SWTBotJunit4ClassRunner;
 import org.eclipse.swtbot.swt.finder.waits.DefaultCondition;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotButton;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotText;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.moreunit.JavaProjectSWTBotTestHelper;
 import org.moreunit.test.context.Context;
 import org.moreunit.test.context.Preferences;
@@ -18,6 +20,7 @@ import org.moreunit.test.context.TestType;
  * @author gianasista
  */
 @Context(mainCls = "org:SomeClass")
+@RunWith(SWTBotJunit4ClassRunner.class)
 public class RenamePackageTest extends JavaProjectSWTBotTestHelper 
 {
 	/*
